backend/requirements.txt
icalevents
Flask
Flask-cors
python-dotenv